refactor: move Vary class to new vary.py
